% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gooogle_main.R
\name{gooogle}
\alias{gooogle}
\title{Group Regularization for zero Inflated Count Regression Models}
\usage{
gooogle(data, xvars, zvars, yvar, group = 1:ncol(data), samegrp.overlap = T,
  penalty = c("grLasso", "grMCP", "grSCAD", "gel", "cMCP", "gBridge",
  "gLasso", "gMCP", "gBridge", "SGL"), dist = c("poisson", "negbin"),
  nlambda = ifelse(penalty == "SGL", 20, 100), lambda,
  lambda.min = ifelse(penalty == "SGL", 0.1, ifelse(nrow(data[,
  unique(c(xvars, zvars))]) > ncol(data[, unique(c(xvars, zvars))])), 1e-04,
  0.05), lambda.max, crit = "BIC", like = "normal", alpha = ifelse(penalty
  == "SGL", 0.95, 1), eps = 0.001, max.iter = 1000,
  gmax = length(unique(group)), gamma = ifelse(penalty == "SGL", 0.8,
  ifelse(penalty == "gBridge", 0.5, ifelse(penalty == "grSCAD", 4, 3))),
  tau = 1/3, warn = TRUE, delta = 1e-07, thresh = 0.001,
  standardize = T, verbose = FALSE, step = 1, reset = 10, ...)
}
\arguments{
\item{data}{This is the data frame containing outcome and predictors}

\item{xvars}{This is the vector of variable names for predictors in count model}

\item{zvars}{This is the vector of variable names for predictors in zero model}

\item{yvar}{This is the outcome variable name}

\item{group}{This is a vector of integers describing the grouping of the coefficients. For greatest efficiency and least ambiguity, it is best if group is a vector of consecutive integers. If any coefficients are to be included in the model without being penalized, their grouping index should be zero}

\item{samegrp.overlap}{This is a logical argument. If TRUE same grouping indices will be assigned to shared predictors in count and zero model}

\item{penalty}{This is the penalty to be applied in the model. For group selection, it is one of grLasso, grMCP, grSCAD or SGL while for bi-level selection it is gBridge or cMCP}

\item{dist}{This is the distribution for count model (Poisson or negative Binomial)}

\item{nlambda}{This is the number of lambda values. Default is 20 for SGL and 100 for other penalties.}

\item{lambda}{This is a user specified sequence of lambda values. By default it is left unspecified and the function automatically computes a grid of lambda values that ranges uniformly on the log scale over a relevant range of the values}

\item{lambda.min}{This is the smallest value of lambda, as a fraction of lambda.max. Default is .0001 if the number of observations is larger than the number of covariates and .05 otherwise.}

\item{lambda.max}{This is the maximum value of lambda (needed for gBridge penalty only). Unlike the penalties in grpreg, it is not possible to solve for lambda.max directly with group bridge models. Thus, it must be specified by the user. If it is not specified, gBridge will attempt a guess at lambda.max which mihgt not be very accurate.}

\item{crit}{This is the selection criteria for the best model. It can be either "BIC" (default) or "AIC"}

\item{alpha}{This is the tuning parameter for the balance between the group penalty and the L2 penalty, as in grpreg. Default is 0.95 for SGL or 1 for other penalties}

\item{eps}{This is the convergence threshhold, as in grpreg.}

\item{gamma}{This is the tuning parameter of the group penalty (see details). Default is 0.8 for SGl, 0.5 for gBridge, 4 for SCAD and 3 for the rest}

\item{tau}{This is the tuning parameter for the group exponential lasso (GEL); defaults to 1/3}

\item{warn}{This is the logical argument which if specified as TRUE gives warning when the function fails to converge as in grpreg. DEFAULT is TRUE}

\item{delta}{This is required for the group bridge penalty only when it is not differentiable at zero, to bound it away from zero. There is typically no need to change this value}

\item{standardize}{This is a logical flag for variable standardization prior to fitting the model}

\item{verbose}{This is the a    logical flag for whether or not step number will be output}

\item{step}{This is the fitting parameter used for inital backtracking step size (between 0 and 1)}

\item{reset}{This is the fitting parameter used for taking advantage of local strong convexity in nesterov momentum (number of iterations before momentum term is reset)}
}
\value{
The function returns a list containing following outputs

coefficients A list containing the estimated coefficients for the count and the zero model

aic The AIC for the fitted model

bic The BIC for the fitted model

loglik The log likelihood for the fitted model
}
\description{
Group Regularization for zero Inflated Count Regression Models
}
\examples{
## Auto Insurance Claim Data
library(HDtweedie)
data("auto")
y<-auto$y
y<-round(y)
x<-auto$x
data<-cbind.data.frame(y,x)
group=c(rep(1,5),rep(2,7),rep(3,4),rep(4:14,each=3),15:21)
yvar<-names(data)[1]
xvars<-names(data)[-1]
zvars<-xvars

## ZIP regression
fit.poisson<-gooogle(data=data,yvar=yvar,xvars=xvars,zvars=zvars,group=group,samegrp.overlap=T,dist="poisson",penalty="gBridge")
fit.poisson$aic

## ZINB regression
fit.negbin<-gooogle(data=data,yvar=yvar,xvars=xvars,zvars=zvars,group=group,samegrp.overlap=T,dist="negbin",penalty="gBridge")
fit.negbin$aic
}

