% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gooogle.R
\name{makePD}
\alias{makePD}
\title{PD approximation of a matrix.}
\usage{
makePD(mat)
}
\arguments{
\item{mat}{A matrix input.}
}
\value{
The function returns a modified matrix which is pd.
}
\description{
The function uses eigen values and eigen vectors to recreate a pd matrix.
}
