% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gooogle.R
\name{gooogle}
\alias{gooogle}
\title{A group regularized fit to the zero inflated count data.}
\usage{
gooogle(data,xvars,zvars,yvar,group=1:ncol(data),samegrp.overlap=T,penalty=c("grLasso", "grMCP", "grSCAD", "gBridge"),dist=c("poisson","negbin"), nlambda=100, lambda,lambda.min=ifelse((nrow(data[,unique(c(xvars,zvars))])>ncol(data[,unique(c(xvars,zvars))])),1e-4,.05),lambda.max, crit="BIC",alpha=1, eps=.001, max.iter=1000, gmax=length(unique(group)),gamma=ifelse(penalty=="gBridge",0.5,ifelse(penalty == "grSCAD", 4, 3)), warn=TRUE)
}
\arguments{
\item{data}{The data frame or matrix consisting of outcome and predictors.}

\item{xvars}{The vector of variable names to be included in count model.}

\item{zvars}{The vector of variable names for excess zero model.}

\item{yvar}{The outcome variable name.}

\item{group}{The vector of integers describing the grouping of the coefficients. For greatest efficiency and least ambiguity, it is best if group is a vector of consecutive integers. If there are coefficientss to be included in the model without being penalized, assign them to group 0 (or "0").}

\item{samegrp.overlap}{A logical argument. If TRUE (default) same grouping indices will be assigned to shared predictors in the count and degenerate distribution.}

\item{penalty}{The penalty to be applied in the model. For group level selection, one of "grLasso", "grMCP" or "grSCAD". For bi-level selection "gBridge" can be specified.}

\item{dist}{The distribution for count model - "poisson" for poisson or "negbin" for negative binomial.}

\item{nlambda}{The number of lambda values. Default is 100.}

\item{lambda}{A user specified sequence of lambda values.}

\item{lambda.min}{The smallest value for lambda, as a fraction of lambda.max. Default is .0001 if the number of observations is larger than the number of covariates and .05 otherwise.}

\item{lambda.max}{The maximum value for lambda (only needed for gBridge penalty).}

\item{crit}{The selection criteria for the best model. It can either be "AIC" or \code{BIC} (default).}

\item{alpha}{The tuning parameter for the balance between the group penalty and the L2 penalty, as in grpreg. Default value is 1.}

\item{eps}{The convergence threshhold, as in grpreg.}

\item{max.iter}{Maximum number of iterations allowed.}

\item{gmax}{Maximum number of non-zero groups allowed.}

\item{gamma}{Tuning parameter of group MCP/SCAD. Default is 3 for MCP and 4 for SCAD.}

\item{warn}{A logical argument indicating whether this function gives warning in case of convergence issue.}
}
\value{
A list containing the following components is returned
\item{coefficients}{A list with two sets of coefficients corresponding to  count and zero inflation parts of the mixture model.}
\item{aic}{The AIC of the selected model.}
\item{bic}{The BIC of the selected model.}
\item{loglik}{The log-likelihood of the selected model.}
}
\description{
Fit zero inflated count data with a group regularization algorithm.
}
\details{
The algorithm fits zero inflated count data to conduct variable selection in the presence of intrinsic grouping structure in the predictor set. Group wise penalties are considered for both count and zero abundance part of the mixture model where the likelihood is optimized using group level or bi-level co-ordinate descent algorithms.
}
\examples{
\dontrun{
## Auto Insurance Claim Data
library(HDtweedie)
data("auto")
y<-auto$y
y<-round(y)
x<-auto$x
data<-cbind.data.frame(y,x)
group=c(rep(1,5),rep(2,7),rep(3,4),rep(4:14,each=3),15:21)
yvar<-names(data)[1]
xvars<-names(data)[-1]
zvars<-xvars

## ZIP regression
fit.poisson<-gooogle(data=data,yvar=yvar,xvars=xvars,zvars=zvars,group=group,samegrp.overlap=T,dist="poisson",penalty="gBridge")
fit.poisson$aic

## ZINB regression
fit.negbin<-gooogle(data=data,yvar=yvar,xvars=xvars,zvars=zvars,group=group,samegrp.overlap=T,dist="negbin",penalty="gBridge")
fit.negbin$aic
}

}
